+++ lib/src/extensions/string_extension.dart
@@ -8,9 +8,7 @@ extension StringNX on String? {
 
 extension StringX on String {
   /// checks whether string is email or not
-  bool get isEmail => RegExp(
-          r'^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$')
-      .hasMatch(this);
+  bool get isEmail => RegExp(r'^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$').hasMatch(this);
 
   /// getOrDefault
   /// returns default value if blank
@@ -49,8 +47,7 @@ extension StringX on String {
   bool get isValidPhoneNumber {
     if (isBlank) return false;
     if (length > 16 || length < 9) return false;
-    return RegExp(r'^[+]*[(]{0,1}[0-9]{1,4}[)]{0,1}[-\s\./0-9]*$')
-        .hasMatch(this);
+    return RegExp(r'^[+]*[(]{0,1}[0-9]{1,4}[)]{0,1}[-\s\./0-9]*$').hasMatch(this);
   }
 
   /// check given string is numeric or not
@@ -94,8 +91,7 @@ extension StringX on String {
     if (isBlank) {
       return false;
     }
-    var regex = RegExp(
-        r'[(http(s)?):\/\/(www\.)?a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)');
+    var regex = RegExp(r'[(http(s)?):\/\/(www\.)?a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)');
     return regex.hasMatch(this);
   }
 
@@ -196,16 +192,7 @@ extension StringX on String {
 final _imageTypes = ['.jpg', '.jpeg', '.png', '.gif', '.webp', '.bmp', '.tif'];
 
 /// audio extensions
-final _audioTypes = [
-  '.mp3',
-  '.weba',
-  '.3gp',
-  '.3g2',
-  '.aac',
-  '.mid',
-  '.midi',
-  '.wav'
-];
+final _audioTypes = ['.mp3', '.weba', '.3gp', '.3g2', '.aac', '.mid', '.midi', '.wav'];
 
 /// video extensions
 final _videoTypes = ['.mp4', '.avi', '.3gp', '.mpeg', '.ogv', '.3g2'];